+++ Controllers/UsuarioController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using tl2_tp10_2023_castroagustin.Models;
 using tl2_tp10_2023_castroagustin.Repositorios;
+using tl2_tp10_2023_castroagustin.ViewModels;
 
 namespace tl2_tp10_2023_castroagustin.Controllers;
 
@@ -18,48 +19,72 @@ public class UsuarioController : Controller
 
     public IActionResult Index()
     {
-        var usuarios = usuarioRepository.GetAll();
-        return View(usuarios);
+        if (!logueado()) return RedirectToRoute(new { controller = "Login", action = "Index" });
+        if (esAdmin())
+        {
+            var usuarios = usuarioRepository.GetAll();
+            return View(new ListarUsuariosViewModel(usuarios));
+        }
+        else
+        {
+            List<Usuario> usuarios = new List<Usuario>();
+            usuarios.Add(usuarioRepository.Get((int)HttpContext.Session.GetInt32("id")));
+            return View(new ListarUsuariosViewModel(usuarios));
+        }
     }
 
     [HttpGet]
     public IActionResult CreateUser()
     {
-        return View(new Usuario());
+        return View(new CrearUsuarioViewModel());
     }
 
     [HttpPost]
-    public IActionResult CreateUser(Usuario usuario)
+    public IActionResult CreateUser(CrearUsuarioViewModel usuario)
     {
-        usuarioRepository.Create(usuario);
+        usuarioRepository.Create(new Usuario(usuario));
         return RedirectToAction("Index");
     }
 
     [HttpGet]
-    public IActionResult UpdateUser(int idUser)
+    public IActionResult UpdateUser(int id)
     {
-        var user = usuarioRepository.Get(idUser);
-        return View(user);
+        if (!logueado()) return RedirectToRoute(new { controller = "Login", action = "Index" });
+        var user = usuarioRepository.Get(id);
+        return View(new ModificarUsuarioViewModel(user));
     }
 
     [HttpPost]
-    public IActionResult UpdateUser(Usuario usuario)
+    public IActionResult UpdateUser(ModificarUsuarioViewModel usuario)
     {
-        usuarioRepository.Update(usuario.Id, usuario);
+        if (!logueado()) return RedirectToRoute(new { controller = "Login", action = "Index" });
+        usuarioRepository.Update(usuario.Id, new Usuario(usuario));
         return RedirectToAction("Index");
     }
 
     [HttpGet]
-    public IActionResult DeleteUser(int idUser)
+    public IActionResult DeleteUser(int id)
     {
-        var user = usuarioRepository.Get(idUser);
+        if (!logueado()) return RedirectToRoute(new { controller = "Login", action = "Index" });
+        var user = usuarioRepository.Get(id);
         return View(user);
     }
 
     [HttpPost]
     public IActionResult DeleteUser(Usuario usuario)
     {
+        if (!logueado()) return RedirectToRoute(new { controller = "Login", action = "Index" });
         usuarioRepository.Remove(usuario.Id);
         return RedirectToAction("Index");
+    }
+
+    private bool logueado()
+    {
+        return HttpContext.Session.Keys.Any();
+    }
+
+    private bool esAdmin()
+    {
+        return HttpContext.Session.Keys.Any() && HttpContext.Session.GetString("rol") == Roles.administrador.ToString();
     }
 }
